Update wayland.xml to version 1.16.0
Now that we've stopped including the system libwayland protocol headers, i.e.
wayland-client.h and wayland-server.h, it should be safe to update the protocol
definitions to the latest version.

Note that this just updates the protocol definition, the implementations for
the various interfaces are still using the old versions.

Change-Id: I961e762553756e11e4c5b33c74101f0c16661232